+++ framework/src/act/build_plan.py
@@ -19,8 +19,7 @@ import pyjson5
 
 from act.build import BuildTask, PythonAction, SubprocessAction, SymlinkAction
 from act.config import CompilerType, Config
-from act.coverreport import generate_report
-from act.merge_summaries import merge_summaries
+from act.coverreport import generate_report, merge_summaries
 from act.parse_test_constraints import TestMetadata
 from act.sail_to_rvvi import sailLog2Trace
 from act.sig_modify import process_signature_file
@@ -426,7 +425,7 @@ def gen_coverage_tasks(
             BuildTask(
                 outputs=(overall_summary,),
                 deps=report_deps,
-                action=PythonAction(fn=merge_summaries, args=(sorted(coverage_reports), overall_summary, config_name)),
+                action=PythonAction(fn=merge_summaries, args=(sorted(coverage_reports), overall_summary)),
             )
         )
